Moved util functions to its own file
